+++ src/main/java/seedu/duke/exceptions/DukeException.java
@@ -6,6 +6,10 @@ public class DukeException extends Exception {
         super();
     }
 
+    public  DukeException(String message) {
+        super(message);
+    }
+
     public DukeException(String message, Throwable cause) {
         super(message, cause);
     }
