+++ src/main/java/seedu/duke/exceptions/InvalidArgumentException.java
@@ -3,7 +3,11 @@ package seedu.duke.exceptions;
 public class InvalidArgumentException extends DukeException {
 
     public InvalidArgumentException() {
+        super();
+    }
 
+    public InvalidArgumentException(String message) {
+        super(message);
     }
 
     public InvalidArgumentException(String message, Throwable cause) {
